last revisi

+++ app/Http/Controllers/BiographyPelatih.php
@@ -254,51 +254,53 @@ class BiographyPelatih extends Controller
         $updated_data = [];
         // dd($gambarSertifikat);
 
+if($gambarSertifikat){
 
-        if (count($gambarSertifikat) > 0) {
+    if (count($gambarSertifikat) > 0) {
 
-            for ($i = 0; $i < count($gambarSertifikat); $i++) {
+        for ($i = 0; $i < count($gambarSertifikat); $i++) {
 
-                $sertifikat = $request->sertifikat[$i];
-                $foto_sertifikat = $gambarSertifikat[$i];
+            $sertifikat = $request->sertifikat[$i];
+            $foto_sertifikat = $gambarSertifikat[$i];
 
-                // Cek apakah data sebelumnya ada atau tidak
-                if (isset($biography->sertifikat[$i])) {
+            // Cek apakah data sebelumnya ada atau tidak
+            if (isset($biography->sertifikat[$i])) {
 
-                    // Cek apakah data diubah atau tidak
-                    if ($biography->sertifikat[$i]['gambar'] == $foto_sertifikat['gambar']) {
-                        // Sama
-                        array_push($updated_data, $foto_sertifikat);
-                    } else {
-                        // Jika Berubah maka hapus file path lama
-                        File::delete(public_path($biography->sertifikat[$i]['gambar']));
-
-                        // Upload file yang berubah
-                        $nama_foto_sertifikat = 'sertifikat_pelatih/binawa_foto_sertifikat_' . $i . '_' . date('Ymdhis') . '.' . $foto_sertifikat['gambar']->getClientOriginalExtension();
-                        $foto_sertifikat['gambar']->move('sertifikat_pelatih', $nama_foto_sertifikat);
-                        $sertifikat['gambar'] = $nama_foto_sertifikat;
-                        array_push($updated_data, $sertifikat);
-                    }
-                    // dd("ada");
+                // Cek apakah data diubah atau tidak
+                if ($biography->sertifikat[$i]['gambar'] == $foto_sertifikat['gambar']) {
+                    // Sama
+                    array_push($updated_data, $foto_sertifikat);
                 } else {
-                    // dd("Tidak Ada");
-                    // Tidak ada
-                    // Upload File Baru
+                    // Jika Berubah maka hapus file path lama
+                    File::delete(public_path($biography->sertifikat[$i]['gambar']));
+
+                    // Upload file yang berubah
                     $nama_foto_sertifikat = 'sertifikat_pelatih/binawa_foto_sertifikat_' . $i . '_' . date('Ymdhis') . '.' . $foto_sertifikat['gambar']->getClientOriginalExtension();
                     $foto_sertifikat['gambar']->move('sertifikat_pelatih', $nama_foto_sertifikat);
                     $sertifikat['gambar'] = $nama_foto_sertifikat;
                     array_push($updated_data, $sertifikat);
                 }
-            }
-            $validatedData['sertifikat'] = $updated_data;
-        } else {
-            // Kosong
-            // dd("Kosong");
-            foreach ($biography->sertifikat as $sertifikat) {
-                // Hapus file
-                File::delete(public_path($sertifikat['gambar']));
+                // dd("ada");
+            } else {
+                // dd("Tidak Ada");
+                // Tidak ada
+                // Upload File Baru
+                $nama_foto_sertifikat = 'sertifikat_pelatih/binawa_foto_sertifikat_' . $i . '_' . date('Ymdhis') . '.' . $foto_sertifikat['gambar']->getClientOriginalExtension();
+                $foto_sertifikat['gambar']->move('sertifikat_pelatih', $nama_foto_sertifikat);
+                $sertifikat['gambar'] = $nama_foto_sertifikat;
+                array_push($updated_data, $sertifikat);
             }
         }
+        $validatedData['sertifikat'] = $updated_data;
+    } else {
+        // Kosong
+        // dd("Kosong");
+        foreach ($biography->sertifikat as $sertifikat) {
+            // Hapus file
+            File::delete(public_path($sertifikat['gambar']));
+        }
+    }
+}
 
 
         // dd($validatedData);
